split SRTNoteBuilder in a separate module
